+++ docs/readme.md
@@ -29,11 +29,11 @@
 
 ### 기능 리스트
 
-- [ ] 주문 받기를 클릭하면 랜덤으로 주문 목록 테이블에 추가됩니다.
-- [ ] 주문 목록의 `수정하기`를 클릭하면 해당 주문을 수정할 수 있습니다.
-- [ ] 주문 목록의 `삭제하기`를 클릭하면 해당 주문이 삭제됩니다.
-- [ ] 주문이 존재하지 않으면 주방의 `커피 선택 버튼`이 `비활성` 됩니다.
-- [ ] 주문이 존재하지 않음에도 `커피 선택 버튼`이나 `만들기` 버튼을 클릭하면 `경고창`을 출력합니다.
+- [x] 주문 받기를 클릭하면 랜덤으로 주문 목록 테이블에 추가됩니다.
+- [x] 주문 목록의 `수정하기`를 클릭하면 해당 주문을 수정할 수 있습니다.
+- [x] 주문 목록의 `삭제하기`를 클릭하면 해당 주문이 삭제됩니다.
+- [x] 주문이 존재하지 않으면 주방의 `커피 선택 버튼`이 `비활성` 됩니다.
+- [x] 주문이 존재하지 않음에도 `커피 선택 버튼`이나 `만들기` 버튼을 클릭하면 `경고창`을 출력합니다.
   
 ### ERD
 ![ERD](source/erd.jpg)
